requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.9.4
ansi2html==1.9.0rc1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
beniget==0.4.1
bleach==6.0.0
blinker==1.7.0
Brotli==1.0.9
category-encoders==2.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
cliapp==1.20180812.1
click==8.1.3
cloudpickle==2.2.1
cmdtest==0.32+git
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
databricks-cli==0.17.4
dataclasses-json==0.5.14
datasets==2.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dglke==0.1.2
dill==0.3.6
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
Django==4.2.5
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
evaluate==0.4.0
executing==1.2.0
f==0.0.1
fastjsonschema==2.16.2
filelock==3.12.3
Flask==3.0.2
Flask-Cors==3.0.10
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fs==2.4.12
fsspec==2023.9.0
gast==0.5.2
gitdb==4.0.10
GitPython==3.1.31
graphdatascience==1.6
greenlet==2.0.2
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==5.2.0
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.16
kaleido==0.2.1
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.0.149
langchain-community==0.0.21
langsmith==0.1.3
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightgbm==3.3.5
llvmlite==0.39.1
loralib==0.1.1
lxml==4.8.0
lz4==3.1.3+dfsg
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.20.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
meson==0.61.2
mistune==2.0.4
mlflow==2.1.1
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
neo4j==5.2.1
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
notebook_shim==0.2.2
npm==0.1.1
numba==0.56.4
numexpr==2.9.0
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
optional-django==0.1.0
orjson==3.8.6
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.13.0
plotly-resampler==0.8.4rc1
ply==3.11
pmdarima==2.0.2
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycaret==3.0.0rc9
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pyod==1.0.7
pyparsing==2.4.7
pypdf==4.0.2
PyPDF2==3.0.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-json-logger==2.0.4
python-multipart==0.0.9
python-slugify==8.0.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
regex==2023.8.8
requests==2.28.2
responses==0.18.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-score==0.1.2
safetensors==0.4.2
schemdraw==0.15
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
sktime==0.16.1
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.9
systemd-python==234
tabulate==0.9.0
tbats==1.1.2
tenacity==8.2.1
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.1.0
tiktoken==0.6.0
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.1
torch==2.2.0
torchdata==0.5.1
torchvision==0.17.0
tornado==6.2
tqdm==4.66.1
trace-updater==0.0.9
traitlets==5.9.0
transformers==4.37.2
triton==2.2.0
ttystatus==0.38
typing-inspect==0.9.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.27.1
virtualenv==20.24.5
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==3.0.1
widgetsnbextension==4.0.5
wrapt==1.15.0rc1
xxhash==3.2.0
yarl==1.9.2
yellowbrick==1.5
zipp==1.0.0
